+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: cfatrane <cfatrane@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/22 10:45:03 by cfatrane          #+#    #+#             */
-/*   Updated: 2016/12/26 18:40:13 by cfatrane         ###   ########.fr       */
+/*   Updated: 2016/12/27 15:31:08 by cfatrane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,8 @@
 # include <stdlib.h>
 # include <string.h>
 # include <unistd.h>
+
+#include <stdio.h>
 
 # define BUFF_SIZE 32
 
@@ -110,12 +112,12 @@ unsigned char		ft_swap_bits(unsigned char octet);
 size_t				ft_tabintlen(int *tab);
 int					**ft_createtab(int nblin, int nbcol);
 int					ft_nbrlen(int nbr);
-int					ft_nbrlen_uns(unsigned long int nb);
+int					ft_nbrlen_uns(unsigned long long int nb);
 int					ft_nbcmp(int nb1, int nb2);
 size_t				ft_lstlen(t_list *list);
 int					ft_count_itoa(int num);
 void				ft_putnbr_base(size_t nb, char *base);
-void				ft_putnbr_uns(long int nb);
+void				ft_putnbr_uns(unsigned long long int nb);
 char				*ft_itoa_base(int num, int base);
 int					ft_size_tab_base(size_t num, int base);
 int					ft_count_words_sep(char const *s, char c);
